Use `macos` folder as source folder
When calling the dropDMG command line tool the `macos` folder is now the source. Previously it was the application bundle inside of the `macos` folder. 

In addition, a `volume name` setting has been added. If not set then the application bundle name minus the `.app` extension will be used as the volume name.